Improve authentication explanation
I was reading through the new authentication docs and I noticed that one
of the use cases for it doesn't really make sense (using a script), and another one I
don't quite understand but I also think it doesn't make sense (testing
an API).

In the end I thought it would be useful to reword this whole section a
bit to explain that we have essentially two ways of getting a token and
how they differ.